Get time for M = 1

+++ Assignment2/dct/dct.h
@@ -2,4 +2,4 @@
 #include <stdlib.h>
 #include <math.h>
 
-void DCTII(int, int, float *, const float *);
+void DCTII(size_t, size_t, float *, const float *);
